techDebt(buRefactorTraining): Move createWorkerKey into shared utils file
