+++ server/routes/establishments/bulkUpload/validate/shared/utils.js
@@ -1,0 +1,3 @@
+exports.createWorkerKey = (localEstablishmentId, workerId) => {
+  return ((localEstablishmentId || '') + (workerId || '')).replace(/\s/g, '');
+};
